이성윤
깃오류 해결...

+++ kh7semimain/src/library/beans/RecommendDto.java
@@ -3,7 +3,7 @@ package library.beans;
 public class RecommendDto {
 	private int recommendNo;
 	private int clientNo;
-	private int bookIsbn;
+	private long bookIsbn;
 	public int getRecommendNo() {
 		return recommendNo;
 	}
@@ -16,10 +16,10 @@ public class RecommendDto {
 	public void setClientNo(int clientNo) {
 		this.clientNo = clientNo;
 	}
-	public int getBookIsbn() {
+	public long getBookIsbn() {
 		return bookIsbn;
 	}
-	public void setBookIsbn(int bookIsbn) {
+	public void setBookIsbn(long bookIsbn) {
 		this.bookIsbn = bookIsbn;
 	}
 	public RecommendDto() {
